Add "showA()" example from Chapter 11 (need to do chapter 11)
This test shows that, as of the current state of the repo, the environment
chain walk causes the showA() function to resolve a to the version defined
on line 8 when it is called on line 9, even though this really doesn't make
sense from a semantics point of view (it "feels" like it should resolve to
the version defined on line 1). That is, we want to implement lexical scope.