Fix closure semantics in calls
I was passing the *current* scope as the closure, but the whole point
of the closure is that it is the scope that existed when the function
was defined. So, this fixes that, which means that the way the calls
function has been changed into a jay_call() function.

It could definitely be interesting to write some optimizations to
turn some calls into much more normal operations.