Study the impact of some possible optimizations
The main target for optimizations that I see is moving things out of
the stack machine and back into local expressions. This can be done
whenever an expression consists entirely of:
1. Arithmetic operations
2. Local variables
3. Number literals

It also seems quite possible that jay_op_call_direct can be used:
essentially, when we know that a VarRef points to a function,
we can directly call the function instead of unwrapping it. Similar
logic will likely apply to classes, etc.