Implement constant folding + associated error messages
This is a fun little optimization that causes expressions entirely
made out of constants to be folded into more constants. This is a
good idea because there's very little reason to believe the C compiler
can tell that these operations can be folded just based on the generated
code, and so should make it faster. It should also help out if any
other optimizations get implemented, e.g. constant propogation.

Finally, the fun thing about this is that it makes it possible to
have more compile-time errors, when we know that an operand is definitely
wrong.

Also, one further direction to take this is to create operators that
have one constant operand, as additional "bytecode" operators. Those should
also be faster than trying to do this like this.