Finalize most of the class setup by adding the dispatcher
The dispatcher is basically a kind of vtable. The nice thing about
it is that it can easily handle the dynamic superclass idea. It would
in some ways be ideal if the methods could be looked up better / devirtualized.

Actually, that is a possible optimization -- on several levels.
First, it is possible to turn a method lookup through the vtable
into a simple index into the methods array. And second, it should
be possible to completely devirtualize a method in some cases,
inlining it into a simple function call. This is the same with
normal functions. With normal functions it's a bit easier to do,
as they can't have naming clashes with other classes, so to speak.