Implement name mangling -- Finish "compilability"?
This should be the last thing needed to make it possible to compile
any Lox program to C.

Of course, they will leak like crazy, as we don't have the garbage
collector. But, it should be enough to call it a compiler, unless
there are still bugs (which is very possible).