Start designing an IR for better temporaries elision
The idea with this IR is that it should make it easier and more generic
to:
- Elide fences
- Elide push/pop pairs

E.g. it should be possible for a = b = c to generate a = b = c in
C code to implement this efficiently.

Also, it may allow for the generation of "temporary double" locations,
such as double tmp1 = JAY_AS_NUMBER(a) - JAY_AS_NUMBER(b); Which will
then be able to be used in expressions.