Add some Val handling for if; make variables much simpler
It seemed somewhat obvious in retrospect: Just have a Val::Variable(VarRef).
This does mean there is no real hope that we can add a std::fmt::Display
for Val that makes the compile code much simpler, but on the other
hand, it does mean that now closures and such work automatically.