Optimization: JAY_MAX_FIELD.
The idea here is that, by looking at all the Set expressions at
compile time, we can figure out exactly which field names might
be in the hash table.

And, as such, we can easily produce a NAME_ for each of those names
first, and then know as a fact that if (name < (last produced name + 1)),
then it is possibly a field, and otherwise it is definitely NOT a
field. This speeds up invoke() by a lot, because it doesn't have to
handle that possibility.