Add an example for what will be needed to optimize addition
Optimizing addition is much harder than the other binary ops because
the operands can be a string or a number. If we can propagate
type information forward, we can optimize a lot. But, we can also
optimize at least a little bit by propagating val information forward
(in the case of the "2 + var" example).